+++ src/main/kotlin/com/github/ferdistro/springpostmanrequestgenerator/annotation/AnnotationData.kt
@@ -1,0 +1,7 @@
+package com.github.ferdistro.springpostmanrequestgenerator.annotation
+
+data class AnnotationData(
+    val name: String,
+    val method: String,
+    val path: String
+)
